moved modules around so it conforms with the requirements
