--- src/modules/csv/testlog.txt
@@ -1,2 +1,0 @@
-Not List or Dictionary
-Not csv file
