+++ src/modules/run_all_tests.bat
@@ -1,2 +1,14 @@
-python3 commands_test.py && cd csv && python3 reader_test.py && python3 writer_test.py && cd ../data && python3 search_test.py && python3 sort_test.py && python3 test_filter.py && cd ../distance_between && python3 geodist_test.py && cd ../ui && python3 test_cli.py && cd ../
+python3 commands_test.py -v 
+cd fileread
+python3 reader_test.py -v 
+cd ../filewrite
+python3 writer_test.py -v 
+cd ../data 
+python3 search_test.py -v 
+python3 sort_test.py -v 
+python3 test_filter.py -v 
+cd ../geodist 
+python3 geodist_test.py -v 
+cd ../ui 
+python3 test_cli.py -v 
 cd ../
